feat: Implement GitHub OAuth authentication and refactor core application setup
